refactor: refactor login component, replace input fields with angular materials, add snackbars for responses
