--- src/app/core/authentication/login/components/login.component.html
@@ -1,40 +1,0 @@
-<div class="login-container">
-  <img src="../../../../assets/images/login-image.jpg" alt="login-image" />
-  <div class="login-form">
-    <h1>Log In</h1>
-    <p>Enter your details below</p>
-    <form (ngSubmit)="onLogin()">
-      <div class="form-group">
-        <input
-          type="text"
-          id="username"
-          name="username"
-          [(ngModel)]="userLogin.username"
-          placeholder="Username"
-          class="input-field"
-        />
-      </div>
-      <div class="form-group">
-        <input
-          type="password"
-          id="password"
-          name="password"
-          [(ngModel)]="userLogin.password"
-          placeholder="Password"
-          class="input-field"
-        />
-      </div>
-      <div class="login-form-actions">
-        <button type="submit" class="login-button">Log In</button>
-        <button type="button" class="gmail-button">
-          <img
-            src="../../../../../assets/images/gmail-icon.png"
-            alt="gmail-icon"
-          />
-          Login with Gmail
-        </button>
-        <p>Don't have an account? <a [routerLink]="['/signup']">Sign up</a></p>
-      </div>
-    </form>
-  </div>
-</div>
