+++ src/app/core/authentication/login/login.component.html
@@ -1,0 +1,55 @@
+<div class="login-container">
+  <img src="../../../assets/images/login-image.jpg" alt="login-image" />
+  <div class="login-form">
+    <h1>Log In</h1>
+    <p>Enter your details below</p>
+    <form [formGroup]="loginForm" (ngSubmit)="onLogin()">
+      <div class="form-group">
+        <mat-form-field appearance="fill">
+          <input
+            matInput
+            type="text"
+            id="username"
+            formControlName="username"
+            placeholder="Username"
+          />
+          <mat-error
+            *ngIf="
+              loginForm.get('username')?.hasError('required') &&
+              loginForm.get('username')?.touched
+            "
+          >
+            Username is required
+          </mat-error>
+        </mat-form-field>
+      </div>
+      <div class="form-group">
+        <mat-form-field appearance="fill">
+          <input
+            matInput
+            type="password"
+            id="password"
+            formControlName="password"
+            placeholder="Password"
+          />
+          <mat-error
+            *ngIf="
+              loginForm.get('password')?.hasError('required') &&
+              loginForm.get('password')?.touched
+            "
+          >
+            Password is required
+          </mat-error>
+        </mat-form-field>
+      </div>
+      <div class="login-form-actions">
+        <button type="submit" class="login-button">Log In</button>
+        <button type="button" class="gmail-button">
+          <img src="../../../assets/images/gmail-icon.png" alt="gmail-icon" />
+          Login with Gmail
+        </button>
+        <p>Don't have an account? <a [routerLink]="['/signup']">Sign up</a></p>
+      </div>
+    </form>
+  </div>
+</div>
