+++ src/app/models/user.model.ts
@@ -2,11 +2,6 @@ export interface IUserLogin {
   username: string;
   password: string;
 }
-export interface ISignupRequest {
-  username: string;
-  email: string;
-  password: string;
-}
 
 export interface IUser {
   id: number;
